Revert changes made by mistake
