--- public/app.module.ts
@@ -1,6 +1,0 @@
-import { NgModule } from '@angular/core';
-import { BrowserModule } from '@angular/platform-browser';
-@NgModule({
-    imports: [ BrowserModule ]
-})
-export default class AppModule {}
